piSerial: Add option -c to specify the device node for encrypt chip
To have a better compatiable with the further tpm option, option -c
is added to support specifying the device node for encrypt chip.

In the meantime,with the option -c, piSerial can deal with the
situation that the device node is changed from the default /dev/i2c-1
to others.

Signed-off-by: Zhi Han <z.han@kunbus.com>