piSerial: remove the redundant calling of get_sn_i2c()
the get_sn_i2c() read the serial number from crypto chip through i2c
and store the serial number to a global variable. Later this serial
number is used for other purposes. Once the serial number is stored
in the global variable, it is not needed to be done again.

Signed-off-by: Zhi Han <z.han@kunbus.com>