piSerial: revpi-factory-reset add support the tpm
it is possible that the tpm of flat can not be accessed because of the missing
of the device tree for it. In this case we just ignore the operation related to
serial number. If/when the tpm chip is available (by checking the existence of
/dev/tpm0), read the serial number and excute the related operation e.g set
passowrd, set machine id...

Signed-off-by: Zhi Han <z.han@kunbus.com>