Send sleep sequence after command transmission
A watchdog counter starts after the ATECC508A has been woken up. After
expiration the device goes back to sleep regardless of whether an I/O
transmission or command execution is in progress.
If this happens while the serial number is read (e.g if the number is read
several consequtive times) the read number may be corrupted.
Prevent this by forcing the device back to sleep after we are done with
reading, since each time the device is woken up, the watchdog counter is
reset.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>
Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>