factory-reset: Fix wrong mac address configuration
Amongst other things revpi-factory-reset sets up the MAC addresses for the
(wireless) network interfaces. For this purpose the MAC address passed to
the script is used to derive different addresses for each network
interface. While the addresses are added correctly to config.txt which is
evaluated at bootup, the addresses set during runtime of the
revpi-factory-reset script are wrong: Instead of different (derived)
addresses only the address passed to the script is used for the interfaces.
This may lead to significant network connection problems as long as the
system is not restarted (and thus the addresses are assigned correctly
during bootup).

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>