factory-reset:flat: Setup MAC address for wireless network interface
Add MAC address setup for the wireless network interface on the flat by
adding the required entries to config.txt and setting the MAC address in
the running system. Remove entries for wireless LAN MAC address beforehand
if such an entry already exits.
Furthermore reformat one if-case for the compact to have a more consistent
layout with the new code.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>