Change the format of mac for "ip link set"
The command "ip link set" accepts only the mac address in the
format of "xx:xx:xx:xx:xx:xx" (case unsensitive), it was in the
format of "xxxxxxxxxxxx" for ifconfig.

As we shifted to tool "ip", it is needed to be changed accordingly.

(FYI: the "ip link set" is only executed in real terminal login
 environment, not in ssh login.)

Signed-off-by: Zhi Han <z.han@kunbus.com>