Fix comparison between string and number
Numbers should be compared for equality with "-eq" while strings are to be
compared using "=", if the comparison is done with [test]. In bash you also
can compare strings using "==", but then should use double brackets for
bash builtin test. In any case comparing string and number this way was
error prone and at least this change makes it consistent with the rest of
the script.
It also fixes a sporadic problem at CM1 introduced by the last commit.

Signed-off-by: Holger Jorra <h.jorra@kunbus.com>