Remove compatibility checks for older setups
In one of the first images (Jessie), the dtoverlay was shared by RevPi
Core and Connect. It was configured statically within the image and
not set by factory-reset. In order to deal with these older setups
compatibility checks where introduced.

Since we are now shipping Buster (two versions after Jessie), this
compatibility fix is not longer needed and will be removed.

Signed-off-by: Zhi Han <z.han@kunbus.com>